refactor: extract pinMode and add tests
