+++ popup.html
@@ -371,6 +371,7 @@
     </form>
   </div>
 
+  <script src="pin.js"></script>
   <script src="popup.js"></script>
   
   <div class="donate-container" style="text-align: center; margin-top: 20px; padding-top: 15px; border-top: 1px solid #eee;">
